+++ src-tauri/Cargo.toml
@@ -45,6 +45,7 @@ chrono = { version = "0.4.38", features = ["serde"] }
 calamine = "0.26.1"
 csv = "1.3.1"
 tauri-plugin-sql = { version = "2", features = ["sqlite"] }
+docx-rs = "0.4.17"
 
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
 tauri-plugin-autostart = "2"
